backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.9.0
chromadb==0.4.22
reportlab==4.0.8
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4